Fix reconnect saved-entry reroute edge cases
