+++ apps/mobile/test/integration/auto-connect-attempt-test-helpers.ts
@@ -1,0 +1,149 @@
+import {
+	type AutoConnectAttemptSourceArgs,
+	attemptAutoConnectSource as attemptAutoConnectSourceBase,
+} from '../../src/lib/auto-connect-attempt';
+import { createConnectionRunContext } from '../../src/lib/connection-run-context';
+import { type SavedConnectionEntry } from '../../src/lib/connection-utils';
+import type { InputConnectionDetails } from '../../src/lib/secrets-manager';
+
+export type OpenSavedEntryShellArgs = {
+	connectionDetails: InputConnectionDetails;
+	resolvedSecurity: {
+		type: 'key';
+		privateKey: string;
+	};
+	navigate: (params: { connectionId: string; channelId: number }) => void;
+};
+
+export const baseDetails: InputConnectionDetails = {
+	username: 'muly',
+	host: 'host.example',
+	port: 22,
+	useTmux: true,
+	tmuxSessionName: 'main',
+	autoConnect: true,
+	security: { type: 'key', keyId: 'key-1' },
+};
+
+export function createLogger() {
+	const calls: unknown[] = [];
+	return {
+		calls,
+		logger: {
+			info: (...args: unknown[]) => {
+				calls.push(['info', ...args]);
+			},
+			warn: (...args: unknown[]) => {
+				calls.push(['warn', ...args]);
+			},
+		},
+	};
+}
+
+export function createSavedEntry(
+	value: SavedConnectionEntry['value'] = baseDetails,
+): SavedConnectionEntry {
+	return {
+		id: 'saved-1',
+		metadata: {
+			createdAtMs: 1,
+			modifiedAtMs: 2,
+			priority: 0,
+		},
+		value,
+	};
+}
+
+export function createSavedEntryWithId(
+	id: string,
+	value: SavedConnectionEntry['value'],
+): SavedConnectionEntry {
+	return {
+		...createSavedEntry(value),
+		id,
+	};
+}
+
+export function activeConnectionFixture(overrides: {
+	connectionId: string;
+	host: string;
+	connectedAtMs?: number;
+	startShell: AutoConnectAttemptSourceArgs['connections'][string]['startShell'];
+}): AutoConnectAttemptSourceArgs['connections'][string] {
+	return {
+		connectionId: overrides.connectionId,
+		connectionDetails: {
+			...baseDetails,
+			host: overrides.host,
+		},
+		connectedAtMs: overrides.connectedAtMs ?? 10,
+		startShell: overrides.startShell,
+	};
+}
+
+export function createAutoConnectRunContext(callerSignal?: AbortSignal) {
+	return createConnectionRunContext({
+		callerSignal,
+		timeouts: {
+			operationTimeoutMs: 60_000,
+			recoveryTimeoutMs: 60_000,
+			cleanupTimeoutMs: 5_000,
+		},
+	});
+}
+
+export async function attemptAutoConnectSource(
+	args: Omit<AutoConnectAttemptSourceArgs, 'runContext'> & {
+		runContext?: AutoConnectAttemptSourceArgs['runContext'];
+		abortSignal?: AbortSignal;
+	},
+) {
+	const runContext =
+		args.runContext ?? createAutoConnectRunContext(args.abortSignal);
+	try {
+		const { abortSignal: _abortSignal, ...sourceArgs } = args;
+		return await attemptAutoConnectSourceBase({ ...sourceArgs, runContext });
+	} finally {
+		if (!args.runContext) {
+			runContext.finish();
+		}
+	}
+}
+
+export function eventKinds(events: unknown[]) {
+	return events.map((event) => (event as { kind: string }).kind);
+}
+
+export function createTaggedError(message: string, tag: string) {
+	const error = new Error(message) as Error & { tag: string };
+	error.tag = tag;
+	return error;
+}
+
+export const unsupportedRecovery = {
+	ensureReady: async () => ({
+		kind: 'unsupported' as const,
+		attempted: false as const,
+		available: false as const,
+	}),
+	recoverAfterFailure: async () => ({
+		kind: 'nonNetworkFailure' as const,
+		attempted: false as const,
+		networkLikeFailure: false as const,
+		available: true,
+	}),
+};
+
+export const readyRecovery = {
+	ensureReady: async () => ({
+		kind: 'ready' as const,
+		attempted: true as const,
+		available: true as const,
+	}),
+	recoverAfterFailure: async () => ({
+		kind: 'nonNetworkFailure' as const,
+		attempted: false as const,
+		networkLikeFailure: false as const,
+		available: true,
+	}),
+};
